refactor: refactor modules/imports of page genetaring
