+++ styles/_btn.scss
@@ -42,3 +42,5 @@
 .meta {
     flex: 1 0 50px;
 }
+
+.btn.active
